apps/backend/requirements.txt
fastapi
uvicorn
google-genai
supabase==2.23.0
python-dotenv
assemblyai
pyannote.audio
torch
librosa
numpy
python-multipart